refactor: modularize tldraw collaboration
