+++ src/components/ui/tldraw/utils/types.ts
@@ -1,0 +1,3 @@
+import type { createLiveKitBus } from '@/lib/livekit/livekit-bus';
+
+export type LiveKitBus = ReturnType<typeof createLiveKitBus>;
